Increase default window size to 1920x1080, update D3D11 debug
Raised the application's default window resolution to 1920x1080 and updated imgui.ini to reflect new UI layout. Added missing headers in Application.cpp. Commented out D3D11 live object reporting in debug mode and improved swap chain resize logging. Changed debug name length type in D3D11Utils for API compatibility.

+++ Kerberos/src/Platform/D3D11/D3D11Utils.h
@@ -9,7 +9,7 @@ namespace Kerberos::D3D11Utils
         _In_ ID3D11DeviceChild* deviceResource,
         _In_z_ const std::string& debugName)
     {
-        if (FAILED(deviceResource->SetPrivateData(WKPDID_D3DDebugObjectName, debugName.length(), debugName.c_str())))
+        if (FAILED(deviceResource->SetPrivateData(WKPDID_D3DDebugObjectName, static_cast<uint32_t>(debugName.length()), debugName.c_str())))
         {
 			KBR_CORE_ERROR("Failed to set debug name for {0}", debugName);
         }
